[ipv6] Use IPv6 fallback route when scope matching fails
When routing to an IPv6 destination, the lookup process ignores routes
that are out of scope. This means any global unicast address is
unreachable if the link only advertises a ULA prefix, which could be
unwanted when subsequent hops implement address translations, like
NPTv6.

This patch implements fallback routing with best efforts, preferring
the largest IPv6 route scope which is iterated in the first place.

Signed-off-by: y5c4l3 <y5c4l3@proton.me>