[efi] replace efi_snp_load_file with a stub when EFI_DOWNGRADE_UX is reuqested
When EFI_DOWNGRADE_UX is requested, efi_load_file_protocol_guid is
replaced with a dummy one, preventing that protocol from registering,
which makes the function efi_snp_load_file useless. This patch replaces
this function with a stub when EFI_DOWNGRADE_UX is requested, to prevent
dragging in the ipxe() symbol, to prevent further dragging in more
symbols and objects.

Signed-off-by: Miao Wang <shankerwangmiao@gmail.com>