[iscsi] Fix reverse CHAP auth
The code currently failed on reversed authing on LIO target.
The error message are :
  iSCSI %p nefarious target tried to bypass authentication

This error is caused by ipxe incorrectly setting the Transition bit.

Accoding to RFC 7143, the Transition bit should only be set on the
final round of CHAP login request. But ipxe set the Transition bit
at all time.

This patch fix this issue by only setting the Transition bit on the
final round of CHAP login request.

Signed-off-by: Zhaoxuan Zhai <zhai.zhaoxuan@canonical.com>